first pass type parameter.ts
