+++ server/package.json
@@ -55,8 +55,10 @@
     "winston": "1.0.2"
   },
   "devDependencies": {
+    "@types/lru-cache": "^5.1.0",
     "@types/node": "^14.14.39",
     "@types/underscore": "^1.11.1",
+    "@types/valid-url": "^1.0.3",
     "typescript": "^4.2.4"
   }
 }
